--- a/src/WinRT.Runtime/ObjectReference.cs
+++ b/src/WinRT.Runtime/ObjectReference.cs
@@ -450,7 +450,7 @@ namespace WinRT
         private readonly IntPtr _contextToken;
 
         private volatile ConcurrentDictionary<IntPtr, ObjectReference<T>> __cachedContext;
-        private ConcurrentDictionary<IntPtr, ObjectReference<T>> _cachedContext => __cachedContext ?? Make_CachedContext();
+        private ConcurrentDictionary<IntPtr, ObjectReference<T>> CachedContext => __cachedContext ?? Make_CachedContext();
         private ConcurrentDictionary<IntPtr, ObjectReference<T>> Make_CachedContext()
         {
             global::System.Threading.Interlocked.CompareExchange(ref __cachedContext, new(), null);
@@ -458,10 +458,9 @@ namespace WinRT
         }
 
         private volatile AgileReference __agileReference;
-        private AgileReference _agileReference => __agileReference ?? Make_AgileReference();
+        private AgileReference AgileReference => __agileReference ?? Make_AgileReference();
         private AgileReference Make_AgileReference()
         { 
-            // Might need to wrap these into a lambda and call it in the CompareExchange 
             AgileReference agileReference = null; 
             Context.CallInContext(_contextCallbackPtr, _contextToken, InitAgileReference, null); 
             void InitAgileReference() { agileReference = new AgileReference(this); }
@@ -477,40 +476,11 @@ namespace WinRT
         {
             _contextCallbackPtr = contextCallbackPtr;
             _contextToken = contextToken;
-            // _cachedContext = new Lazy<ConcurrentDictionary<IntPtr, ObjectReference<T>>>();
-            /*
-            _agileReference = new Lazy<AgileReference>(() => {
-                AgileReference agileReference = null;
-                Context.CallInContext(_contextCallbackPtr, _contextToken, InitAgileReference, null);
-                return agileReference;
-
-                void InitAgileReference()
-                {
-                    agileReference = new AgileReference(this);
-                }
-            });
-            */
         }
 
         internal ObjectReferenceWithContext(IntPtr thisPtr, IntPtr contextCallbackPtr, IntPtr contextToken, Guid iid)
-            : base(thisPtr)
+            : this(thisPtr, contextCallbackPtr, contextToken)
         {
-            _contextCallbackPtr = contextCallbackPtr;
-            _contextToken = contextToken;
-            // _cachedContext = new Lazy<ConcurrentDictionary<IntPtr, ObjectReference<T>>>();
-            /* 
-            _agileReference = new Lazy<AgileReference>(() => {
-                AgileReference agileReference = null;
-                Context.CallInContext(_contextCallbackPtr, _contextToken, InitAgileReference, null);
-                return agileReference;
-
-                void InitAgileReference()
-                {
-                    agileReference = new AgileReference(this);
-                }
-            });
-            */
-
             _iid = iid;
         }
 
@@ -548,11 +518,11 @@ namespace WinRT
                 return null;
             }
 
-            return _cachedContext.GetOrAdd(currentContext, CreateForCurrentContext);
+            return CachedContext.GetOrAdd(currentContext, CreateForCurrentContext);
 
             ObjectReference<T> CreateForCurrentContext(IntPtr _)
             {
-                var agileReference = _agileReference;
+                var agileReference = AgileReference;
                 // We may fail to switch context and thereby not get an agile reference.
                 // In these cases, fallback to using the current context.
                 if (agileReference == null)
@@ -574,9 +544,9 @@ namespace WinRT
 
         protected override unsafe void Release()
         {
-            if (_cachedContext != null)
+            if (__cachedContext != null)
             {
-                _cachedContext.Clear();
+                CachedContext.Clear();
             }
 
             Context.CallInContext(_contextCallbackPtr, _contextToken, base.Release, ReleaseWithoutContext);