This change brings in Joshua's change https://github.com/microsoft/CsWinRT/commit/3122ce19368d2d593de239f17fb56410ebfd0844 for removing uses of Lazy in ObjectReferenceWithContext and adds some benchmarks and a couple other improvements in agile references which are used in ObjectReferenceWithContext.

Master

|                          Method |     Mean |    Error |   StdDev | Allocated |
|-------------------------------- |---------:|---------:|---------:|----------:|
| ConstructAndQueryNonAgileObject | 15.74 ms | 0.287 ms | 0.341 ms |      8 KB |
|         ConstructNonAgileObject | 15.70 ms | 0.285 ms | 0.292 ms |      3 KB |


PR

|                          Method |     Mean |    Error |   StdDev | Allocated |
|-------------------------------- |---------:|---------:|---------:|----------:|
| ConstructAndQueryNonAgileObject | 15.74 ms | 0.309 ms | 0.289 ms |      5 KB |
|         ConstructNonAgileObject | 15.79 ms | 0.161 ms | 0.151 ms |      2 KB |

Overall the perf difference is within the error range, but we do reduce allocations which can be helpful with reducing GC time spent later.